+++ scripts/LOS/PointSourcePolygon.js
@@ -97,7 +97,7 @@ PATCHES.LOS.METHODS = {
  *                                  If undefined, all walls will be treated as NORMAL restriction.
  * @returns {boolean|object[]|object}
  */
-function testWallsForIntersections(origin, destination, walls, mode, type) {
+export function testWallsForIntersections(origin, destination, walls, mode, type) {
   origin = new Point3d(origin.x, origin.y, origin.z);
   destination = new Point3d(destination.x, destination.y, destination.z);
   const direction = destination.subtract(origin);
@@ -105,7 +105,7 @@ function testWallsForIntersections(origin, destination, walls, mode, type) {
   const collisions = [];
   for ( let wall of walls ) {
     // Check the 2d overhead first.
-    if ( !foundry.utils.lineSegmentIntersects(origin, destination, wall.A, wall.B) ) continue;
+    if ( !foundry.utils.lineSegmentIntersects(origin, destination, wall.edge.a, wall.edge.b) ) continue;
 
     const wallPoints = Point3d.fromWall(wall, { finite: true });
     const t = Plane.rayIntersectionQuad3dLD(
@@ -145,15 +145,15 @@ function originalTestWallInclusion(wall, bounds) {
   const {type, boundaryShapes, useThreshold, wallDirectionMode } = this.config;
 
   // First test for inclusion in our overall bounding box
-  if ( !bounds.lineSegmentIntersects(wall.A, wall.B, { inside: true }) ) return false;
+  if ( !bounds.lineSegmentIntersects(wall.edge.a, wall.edge.b, { inside: true }) ) return false;
 
   // Specific boundary shapes may impose additional requirements
   for ( const shape of boundaryShapes ) {
-    if ( shape._includeEdge && !shape._includeEdge(wall.A, wall.B) ) return false;
+    if ( shape._includeEdge && !shape._includeEdge(wall.edge.a, wall.edge.b) ) return false;
   }
 
   // Ignore walls which are nearly collinear with the origin
-  const side = wall.orientPoint(this.origin);
+  const side = wall.edge.orientPoint(this.origin);
   if ( !side ) return false;
 
   // Always include interior walls underneath active roof tiles